requirements.txt
alembic==1.7.6
Authlib==1.1.0
bcrypt==3.2.0
bidict==0.22.0
blinker==1.4
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.0
charset-normalizer==2.1.1
click==8.0.3
cryptography==36.0.2
DateTime==4.4
dnspython==2.2.0
ecdsa==0.17.0
email-validator==1.1.3
Flask==2.1.1
Flask-Cors==3.0.10
Flask-Executor==0.10.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Moment==1.0.2
Flask-SocketIO==5.3.2
Flask-WTF==1.0.0
google-api-core==2.10.1
google-api-python-client==2.62.0
google-auth==2.11.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.4.0
itsdangerous==2.1.2
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
oauthlib==3.2.1
passlib==1.7.4
peewee==3.14.8
protobuf==4.21.6
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dotenv==0.19.2
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2021.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.31
uritemplate==4.1.1
urllib3==1.26.12
uWSGI==2.0.20
Werkzeug==2.0.2
WTForms==3.0.1
zipp==3.7.0
zope.interface==5.4.0
